Remove NullableSetting.java sample field style
This is an illustration about how builder fields generally work, but it is never
actually used in the annotation processor, and indeed, will always be useless
because this behavior is baked-in.

It's not trivially possible to use this field style for real as the default,
because it automatically boxes primitive types.